+++ src/main/java/com/example/spot/domain/mapping/MemberStudy.java
@@ -1,8 +1,9 @@
 package com.example.spot.domain.mapping;
 
 import com.example.spot.domain.Member;
-import com.example.spot.domain.Study;
+
 import com.example.spot.domain.enums.ApplicationStatus;
+import com.example.spot.domain.study.Study;
 import jakarta.persistence.*;
 import lombok.Getter;
