+++ src/main/java/com/example/spot/domain/mapping/MemberVote.java
@@ -1,13 +1,16 @@
 package com.example.spot.domain.mapping.mapping;
 
-import com.example.domain.common.BaseEntity;
+import com.example.spot.domain.Member;
+import com.example.spot.domain.common.BaseEntity;
 import com.example.spot.domain.study.Option;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -28,8 +31,11 @@ public class MemberVote extends BaseEntity {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
-   // private Member member
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member member;
 
     @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "option_id", nullable = false)
     private Option option;
 }
