+++ src/main/java/com/example/spot/domain/mapping/PreferredRegion.java
@@ -1,7 +1,6 @@
 package com.example.spot.domain.mapping;
 
 import com.example.spot.domain.Member;
-import com.example.spot.domain.Region;
 import jakarta.persistence.*;
 import lombok.Getter;
 
@@ -18,8 +17,8 @@ public class PreferredRegion {
     @JoinColumn(name = "member_id", nullable = false)
     private Member member;
 
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "region_id", nullable = false)
-    private Region region;
+    //@ManyToOne(fetch = FetchType.LAZY)
+    //@JoinColumn(name = "region_id", nullable = false)
+    //private Region region;
 
 }
