+++ src/main/java/com/example/spot/domain/mapping/PreferredStudy.java
@@ -1,7 +1,7 @@
 package com.example.spot.domain.mapping;
 
 import com.example.spot.domain.Member;
-import com.example.spot.domain.Study;
+import com.example.spot.domain.study.Study;
 import jakarta.persistence.*;
 import lombok.Getter;
