+++ src/main/java/com/example/spot/domain/mapping/StudyLikedPost.java
@@ -1,12 +1,14 @@
-package com.example.spot.domain.mapping.mapping;
+package com.example.spot.domain.mapping;
 
-import com.example.domain.common.BaseEntity;
+import com.example.spot.domain.Member;
+import com.example.spot.domain.common.BaseEntity;
 import com.example.spot.domain.study.StudyPost;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -30,8 +32,11 @@ public class StudyLikedPost extends BaseEntity {
     private long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "study_post_id", nullable = false)
     private StudyPost studyPost;
 
-    //private Member member;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member member;
 
 }
