+++ src/main/java/com/example/spot/domain/mapping/StudyPostImage.java
@@ -1,6 +1,6 @@
-package com.example.spot.domain.mapping.mapping;
+package com.example.spot.domain.mapping;
 
-import com.example.domain.common.BaseEntity;
+import com.example.spot.domain.common.BaseEntity;
 import com.example.spot.domain.study.StudyPost;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -8,6 +8,7 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -33,6 +34,7 @@ public class StudyPostImage extends BaseEntity {
     private String url;
 
     @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "study_post_id", nullable = false)
     private StudyPost studyPost;
 
 }
