+++ src/main/java/com/example/spot/domain/mapping/StudyTheme.java
@@ -1,7 +1,7 @@
 package com.example.spot.domain.mapping;
 
-import com.example.spot.domain.Study;
 import com.example.spot.domain.Theme;
+import com.example.spot.domain.study.Study;
 import jakarta.persistence.*;
 import lombok.Getter;
