+++ src/main/java/com/example/spot/domain/study/StudyPostComment.java
@@ -1,13 +1,16 @@
 package com.example.spot.domain.study;
 
-import com.example.domain.common.BaseEntity;
-import com.example.spot.domain.mapping.mapping.StudyLikedComment;
+
+import com.example.spot.domain.Member;
+import com.example.spot.domain.common.BaseEntity;
+import com.example.spot.domain.mapping.StudyLikedComment;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
 import java.util.List;
@@ -34,8 +37,9 @@ public class StudyPostComment extends BaseEntity {
     @ManyToOne(fetch = FetchType.LAZY)
     private StudyPost studyPost;
 
-    // @Column(nullable = false)
-    // private Member member;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Member member;
 
     @Column(nullable = false)
     private String content;
